+++ lib/sendgrid/marketing_campaigns/contacts/lists.ex
@@ -8,15 +8,39 @@ defmodule SendGrid.Contacts.Lists do
 
   @base_api_url "/v3/contactdb/lists"
 
-
   @doc """
   Retrieves all email lists.
   """
   @spec all() :: list(%{}) | :error
   def all() do
-    SendGrid.get(@base_api_url)
-    |> case do
-      { :ok, %{ status_code: 200, body: body } } -> body["lists"]
+    case SendGrid.get(@base_api_url) do
+      {:ok, %{status_code: 200, body: body}} -> body["lists"]
+      _ -> :error
+    end
+  end
+
+  @doc """
+  Creates an email list.
+
+      {:ok, 2} = add("marketing")
+
+  """
+  @spec add(String.t) :: {:ok, integer} | :error
+  def add(list_name) do
+    case SendGrid.post(@base_api_url, %{name: list_name}) do
+      {:ok, %{status_code: 201, body: body}} -> {:ok, body["id"]}
+      _ -> :error
+    end
+  end
+
+  @doc """
+  Retrieves all recipients from an email list.
+  """
+  @spec all_recipients(integer, integer, integer) :: list(%{}) | :error
+  def all_recipients(list_id, page \\ 1, page_size \\ 100) do
+    url = @base_api_url <> "/#{list_id}/recipients?page_size=#{page_size}&page=#{page}"
+    case SendGrid.get(url) do
+      {:ok, %{status_code: 200, body: body}} -> body["recipients"]
       _ -> :error
     end
   end
@@ -25,12 +49,28 @@ defmodule SendGrid.Contacts.Lists do
   Adds a recipient to an email list.
 
       :ok = add_recipient(123, "recipient_id")
+  
   """
   @spec add_recipient(integer, String.t) :: :ok | :error
   def add_recipient(list_id, recipient_id) do
-    SendGrid.post(@base_api_url <> "/#{list_id}/recipients/#{recipient_id}", %{})
-    |> case do
-      { :ok, %{ status_code: 201 } } -> :ok
+    url = @base_api_url <> "/#{list_id}/recipients/#{recipient_id}"
+    case SendGrid.post(url, %{}) do
+      {:ok, %{status_code: 201}} -> :ok
+      _ -> :error
+    end
+  end
+
+  @doc """
+  Deletes a recipient from an email list.
+
+      :ok = delete_recipient(123, "recipient_id")
+
+  """
+  @spec delete_recipient(integer, String.t) :: :ok | :error
+  def delete_recipient(list_id, recipient_id) do
+    url = @base_api_url <> "/#{list_id}/recipients/#{recipient_id}"
+    case SendGrid.delete(url, %{}) do
+      {:ok, %{status_code: 204}} -> :ok
       _ -> :error
     end
   end